Move VideoListItem into its own file
