+++ src/components/AuthDialog/index.tsx
@@ -1,5 +1,4 @@
 import {
-  Box,
   Button,
   CloseButton,
   Dialog,
